its a float now (double only)

+++ Dashboard/views.py
@@ -47,7 +47,7 @@ def AWSUpdate(request):
 
     l = [li]
 
-    if double(luxLevel) < 20:
+    if float(luxLevel) < 20:
         if li.motionDetected or li.overrideMotionSensor:
             li.isOn = True
